// src/app/modules/estimate-billing/utils/estimates.types.ts
import { T_FUEL_TYPE } from '../../cars/cars.types';
import { IAddress } from '../../contacts/contacts.types';

export type IEstimateType = 'SALE' | 'SERVICE';
export interface IEstimate {
    id: string;
    estimateNumber: string;
    estimateDate: number; // Was invoiceDate
    date: number;
    type: IEstimateType;

    billTo: {
        id?: string;
        estimateNumber?: string; // Was invoiceNumber
        name: string;
        phoneNumber: {
            code: string;
            number: string;
        };
        email?: string;
        addressLine1: string;
        addressLine2?: string;
        country: string;
        postalCode: string;
        address?: IAddress;
        postCode: string;
        city: string;
        createdDate: number;
    };

    carInfo?: {
        id?: string;
        regNo: string;
        regYear: string;
        make: string;
        model: string;
        engineType: string;
        transmission: string;
        fuelType: T_FUEL_TYPE;
        mileage: string;
        color: string;
        vin: string;
        nextServiceDate?: string;
        motValidTill?: string;
        insuranceValidTill?: string;
        roadTaxValidTill?: string;
    };

    services: IService[];

    tax: {
        unit: 'percentage' | 'amount';
        value: number;
    };
    discount: {
        unit: 'percentage' | 'amount';
        value: number;
    };
    subtotal?: number;
    total: number;
    advance?: number;
    partExchange?: number;
    balance?: number;

    hasWarranty?: boolean;
}

export type TEstimateTimeFilter =
    | '1m'
    | '3m'
    | '6m'
    | 'cfy'
    | 'lfy'
    | 'dr'
    | ''
    | 'all';
export type TEstimateTypeFilter = 'ALL' | 'SALE' | 'SERVICE' | '';

/** For a SALE bill, the data in the fields are different and mentioned in a comment on the side */
export interface IService {
    id: string; // Make
    item: string; // Model
    quantity: number; // Defaulted to 1
    price: number; // Price
    total: number; // Reg No
}